renamed readhex
